requirements.txt
flask
flask-cors
openai
chromadb
python-dotenv
numpy
pandas
requests
gunicorn
uvicornpy